f32 -> f64

+++ src/color_rgb/color_rgb.rs
@@ -45,44 +45,44 @@ impl Add for ColorRgb {
     }
 }
 
-impl Mul<f32> for ColorRgb {
+impl Mul<f64> for ColorRgb {
     type Output = Self;
-    fn mul(self, mult: f32) -> Self {
+    fn mul(self, mult: f64) -> Self {
         ColorRgb {
-            red: (self.red as f32 * mult) as u8,
-            green: (self.green as f32 * mult) as u8,
-            blue: (self.blue as f32 * mult) as u8,
+            red: (self.red as f64 * mult) as u8,
+            green: (self.green as f64 * mult) as u8,
+            blue: (self.blue as f64 * mult) as u8,
         }
     }
 }
 
-impl MulAssign<f32> for ColorRgb {
-    fn mul_assign(&mut self, mul: f32) {
-        self.red = (self.red as f32 * mul) as u8;
-        self.green = (self.green as f32 * mul) as u8;
-        self.blue = (self.blue as f32 * mul) as u8;
+impl MulAssign<f64> for ColorRgb {
+    fn mul_assign(&mut self, mul: f64) {
+        self.red = (self.red as f64 * mul) as u8;
+        self.green = (self.green as f64 * mul) as u8;
+        self.blue = (self.blue as f64 * mul) as u8;
     }
 }
 
-impl Div<f32> for ColorRgb {
+impl Div<f64> for ColorRgb {
     type Output = Self;
-    fn div(self, div: f32) -> Self {
+    fn div(self, div: f64) -> Self {
         if div == 0.0 {
             panic!("Divide by Zero")
         }
         ColorRgb {
-            red: (self.red as f32 / div) as u8,
-            green: (self.green as f32 / div) as u8,
-            blue: (self.blue as f32 / div) as u8,
+            red: (self.red as f64 / div) as u8,
+            green: (self.green as f64 / div) as u8,
+            blue: (self.blue as f64 / div) as u8,
         }
     }
 }
 
-impl DivAssign<f32> for ColorRgb {
-    fn div_assign(&mut self, div: f32) {
-        self.red = (self.red as f32 / div) as u8;
-        self.green = (self.green as f32 / div) as u8;
-        self.blue = (self.blue as f32 / div) as u8;
+impl DivAssign<f64> for ColorRgb {
+    fn div_assign(&mut self, div: f64) {
+        self.red = (self.red as f64 / div) as u8;
+        self.green = (self.green as f64 / div) as u8;
+        self.blue = (self.blue as f64 / div) as u8;
     }
 }
